+++ src/main/java/tipitapi/drawmytoday/common/entity/BaseEntityWithUpdate.java
@@ -6,14 +6,13 @@ import javax.persistence.EntityListeners;
 import javax.persistence.MappedSuperclass;
 import javax.validation.constraints.NotNull;
 import lombok.Getter;
-import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.annotation.LastModifiedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
 @Getter
 @MappedSuperclass
 @EntityListeners(AuditingEntityListener.class)
-public abstract class BaseEntityWithUpdate extends BaseEntity{
+public abstract class BaseEntityWithUpdate extends BaseEntity {
 
   @NotNull
   @LastModifiedDate
