+++ src/main/java/tipitapi/drawmytoday/common/exception/CustomException.java
@@ -1,12 +1,11 @@
 package tipitapi.drawmytoday.common.exception;
 
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpStatus;
-
 import lombok.Getter;
+import lombok.RequiredArgsConstructor;
 
 @Getter
 @RequiredArgsConstructor
 public class CustomException extends RuntimeException {
+
   private final ErrorCode errorCode;
 }
