+++ src/main/java/tipitapi/drawmytoday/common/util/ParseUserId.java
@@ -4,6 +4,7 @@ import java.util.Optional;
 import org.springframework.security.core.Authentication;
 
 public class ParseUserId {
+
   public static Long parseUserId(Authentication authentication) {
     if (Optional.ofNullable(authentication).isPresent()) {
       return Long.parseLong(authentication.getName());
