+++ src/main/java/tipitapi/drawmytoday/diary/domain/Image.java
@@ -1,6 +1,5 @@
 package tipitapi.drawmytoday.diary.domain;
 
-import java.time.LocalDateTime;
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.FetchType;
@@ -13,7 +12,6 @@ import javax.validation.constraints.NotNull;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import org.springframework.data.annotation.CreatedDate;
 import tipitapi.drawmytoday.common.entity.BaseEntity;
 
 @Getter
